refactor: ModeStartUseCase를 만들어 Input, Output을 가질 수 있도록 수정
