+++ Tempus/Presentation/ClockScene/ViewModel/ClockViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  ClockViewModel.swift
+//  Tempus
+//
+//  Created by 이정민 on 2023/04/23.
+//
+
+import Foundation
